+++ packages/nx/lib/prelude.ml
@@ -3,6 +3,6 @@
   SPDX-License-Identifier: ISC
   ---------------------------------------------------------------------------*)
 
-#install_printer Nx.pp_data;;
+#install_printer Nx.pp;;
 
 open Nx